moved error handling related methods from app to errorhandler
